moved styles to folder styles
