+++ src/styled.tsx
@@ -1,4 +1,6 @@
-import { Theme } from 'colours'
-import styled, { CreateStyled } from "react-emotion";
+import styled, { CreateStyled } from "react-emotion"
+
+import { Theme } from 'styles'
+
 
 export default styled as CreateStyled<Theme>
