+++ src/styles/colours.ts
@@ -1,0 +1,16 @@
+export const colours = {
+  primaryColour: '#7fb3ff',
+  secondaryColour: '#beccdc',
+  badColour: '#e95656',
+  goodColour: '#91e4a5',
+  searchBarColour: '#415f8a',
+  linkHoverColour: '#e0be86',
+  bgStartColour: '#012861',
+  bgMiddleColour: '#021738',
+  bgEndColour: '#021025',
+  bgSearchStartColour: '#001330',
+  bgSearchEndColour: 'rgba(0, 8, 19, 0.8)',
+  separatorColour: '#54606d80'
+}
+
+
