+++ src/styles/index.styles.ts
@@ -1,0 +1,7 @@
+import { css } from "react-emotion"
+
+export const stockTrackerPadding = `1.5rem`
+
+export const logo = css`
+  fill: white;
+`
