+++ src/styles/index.ts
@@ -1,1 +1,6 @@
 import './globals'
+
+export  { Label } from './label.styles'
+export { logo, stockTrackerPadding } from './index.styles'
+export { colours } from './colours'
+export { theme, Theme } from './theme'
